app/requirements.txt
python-dotenv==1.0.1
huggingface_hub
Flask
fastapi
psycopg2
langchain
langchain-community

langchain-core
langchain-experimental
langchain-mistralai
mistralai


